Refactor serialization utility for callnum
