+++ lib/serializers/bib.js
@@ -6,7 +6,7 @@ const Lookup = require('../lookup')
 const buildFieldMapper = require('../field-mapper')
 const formattingUtils = require('./formatting-utils')
 const utils = require('../utils')
-const { getBibCallNum } = require('./serializerUtilities')
+const { getBibCallNum } = require('../serialization-utils')
 const sierraLocationMapping = require('@nypl/nypl-core-objects')('by-sierra-location')
 
 function InvalidInputError (message) {
@@ -354,7 +354,16 @@ var fromMarcJson = (object, datasource) => {
       })
     })
 
-    getBibCallNum(object, builder, fieldMapper)
+    const callnum = getBibCallNum(object)
+    if (callnum && callnum.value) {
+      // Pull callnumber suffix from fieldTag v if present
+      var callnumSuffix = object.fieldTag('v')
+      if (callnumSuffix && callnumSuffix.length) {
+        callnum.value += ' ' + callnumSuffix[0]
+        callnum.path += ', fieldTag v'
+      }
+      builder.add(fieldMapper.predicateFor('Call number'), { literal: callnum.value }, 0, { path: callnum.path })
+    }
 
     // subjectLiteral formatter
     const formatSubjectLiteral = (hash) => {
