+++ lib/serializers/item.js
@@ -5,7 +5,7 @@ const statusMapping = require('@nypl/nypl-core-objects')('by-statuses')
 const catalogItemTypeMapping = require('@nypl/nypl-core-objects')('by-catalog-item-types')
 
 const utils = require('../utils')
-const { getBibCallNum } = require('./serializerUtilities')
+const { getBibCallNum } = require('../serialization-utils')
 const Statement = require('../models/statement')
 const Creator = require('../models/creator')
 const Datasource = require('../models/datasource')
@@ -149,7 +149,7 @@ const fromMarcJson = async (object, datasource, bib) => {
 
     // If not found, default to callnum from bib
     if (!callnum || !callnum.value) {
-      const bibCallNum = getBibCallNum(bib, builder, fieldMapper, false)
+      const bibCallNum = getBibCallNum(bib)
       if (bibCallNum && bibCallNum.value) callnum = bibCallNum
     }
